Prototype for paste and drag drop support for the editor
closes #4753
- added drag drop support for images into the editor

change image name in the image tag to the resulting image name from the /uploads result

- move UploadUi to separate class so it can be reused
- show upload progress when pasting and dropping files in the editor

- add fail event to upload
- add the right extension when pasting image based on the mimeType

change image name in the image tag to the resulting image name from the /uploads result

- move UploadUi to separate class so it can be reused
- show upload progress when pasting and dropping files in the editor

add icon-trash icon back to cancel button after rebase

remove unused Ghost + ghostPaths
add self variable to fix undefined variable warning

jscs warnings
- added space after function
- removed empty line after opening brace
- fixed invalid quote mark

add <i> icon-link to image dropzone

fixed jshint: undefined Ghost variable

changed ghostPaths var to function call

removed double set value call
changed editor api change event to setValue to support firefox change event

changed function prototype to didInsertElement method

changed how event is triggered

moved attachFileHandler to the right didInsertElement

fix rebase errors

added changes from master

fixed wrong dropzone variable
