+++ core/client/app/mixins/ed-editor-api.js
@@ -128,7 +128,8 @@ export default Ember.Mixin.create({
         }
 
         $textarea.focus();
+
         // Tell the editor it has changed, as programmatic replacements won't trigger this automatically
-        this.sendAction('onChange');
+        this.$().trigger('change');
     }
 });
